Add wrapper around badpix from flats and darks.
